requirements.txt
streamlit
pillow
google-generativeai
streamlit-option-mean
ptl